combine mobile and section
