+++ frontend/components/ResultsPage/Results/SearchResult.tsx
@@ -1,7 +1,7 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { Markup } from 'interweave'
 import macros from '../../macros'
-import DesktopSectionPanel from './DesktopSectionPanel'
+import { DesktopSectionPanel, MobileSectionPanel } from './SectionPanel'
 import { Course, PrereqType } from '../../types'
 import IconGlobe from '../../images/IconGlobe'
 import IconArrow from '../../images/IconArrow'
@@ -10,21 +10,23 @@ import useResultDetail from './useResultDetail'
 import useUserChange from './useUserChange';
 import useShowAll from './useShowAll';
 import moment from 'moment';
+import MobileCollapsableDetail from './MobileCollapsableDetail'
+import IconCollapseExpand from '../../images/IconCollapseExpand'
 
 interface SearchResultProps {
   aClass: Course,
 }
-export default function SearchResult({ aClass } : SearchResultProps) {
+
+const getLastUpdateString = (course: Course) : string => {
+  return course.lastUpdateTime ? moment(course.lastUpdateTime).fromNow() : null;
+}
+
+export function SearchResult({ aClass } : SearchResultProps) {
   const { optionalDisplay, creditsString } = useResultDetail(aClass)
   const userIsWatchingClass = useUserChange(aClass)
   const {
     showAll, setShowAll, renderedSections, hideShowAll,
   } = useShowAll(aClass)
-
-  // FIXME move to utils?
-  const getLastUpdateString = (course: Course) : string => {
-    return course.lastUpdateTime ? moment(course.lastUpdateTime).fromNow() : null;
-  }
 
   const feeString = aClass.feeDescription && aClass.feeAmount ? `${aClass.feeDescription}- $${aClass.feeAmount}` : null
 
@@ -105,5 +107,76 @@ export default function SearchResult({ aClass } : SearchResultProps) {
       </div>
       )}
     </div>
+  )
+}
+
+export function MobileSearchResult({ aClass } : SearchResultProps) {
+  const [expanded, setExpanded] = useState(false)
+  const [showMore, setShowMore] = useState(false)
+  const [showNUPath, setShowNUPath] = useState(false)
+  const [showPrereq, setShowPrereq] = useState(false)
+  const [showCoreq, setShowCoreq] = useState(false)
+  const userIsWatchingClass = useUserChange(aClass)
+  const {
+    showAll, setShowAll, renderedSections, hideShowAll,
+  } = useShowAll(aClass)
+
+  const { optionalDisplay, creditsString } = useResultDetail(aClass);
+
+
+  const renderNUPaths = () => (
+    // eslint-disable-next-line react/prop-types
+    <div>{aClass.nupath.length > 0 ? <div> {aClass.nupath.join(', ')}</div> : <span className='empty'> None</span>}</div>
+  )
+
+  return (
+    <div className='MobileSearchResult'>
+      <div className={ expanded ? 'MobileSearchResult__header--expanded' : 'MobileSearchResult__header' } role='button' tabIndex={ 0 } onClick={ () => setExpanded(!expanded) }>
+        <IconCollapseExpand />
+        <span className='MobileSearchResult__header--classTitle'>
+          {`${aClass.subject} ${aClass.classId} : ${aClass.name}`}
+        </span>
+      </div>
+      {expanded && (
+      <div className='MobileSearchResult__panel'>
+        <div className='MobileSearchResult__panel--mainContainer'>
+          <div className='MobileSearchResult__panel--infoStrings'>
+            <a href={ aClass.prettyUrl } target='_blank' rel='noopener noreferrer'>{`Updated ${(getLastUpdateString(aClass))}`}</a>
+            <span>
+              {creditsString()}
+            </span>
+          </div>
+          <div className={ showMore ? 'MobileSearchResult__panel--description' : 'MobileSearchResult__panel--descriptionHidden' }>
+            {aClass.desc}
+          </div>
+          <div className='MobileSearchResult__panel--showMore' role='button' tabIndex={ 0 } onClick={ () => setShowMore(!showMore) }>{showMore ? 'Show less' : 'Show more'}</div>
+          <MobileCollapsableDetail title='NUPATH' expand={ showNUPath } setExpand={ setShowNUPath } renderChildren={ renderNUPaths } />
+          <MobileCollapsableDetail title='PREREQUISITES' expand={ showPrereq } setExpand={ setShowPrereq } renderChildren={ () => optionalDisplay(PrereqType.PREREQ, aClass) } />
+          <MobileCollapsableDetail title='COREQUISITES' expand={ showCoreq } setExpand={ setShowCoreq } renderChildren={ () => optionalDisplay(PrereqType.COREQ, aClass) } />
+          <div className='MobileSearchResult__panel--notifContainer'>
+            <SignUpForNotifications aClass={ aClass } userIsWatchingClass={ userIsWatchingClass } />
+          </div>
+        </div>
+        <div className='MobileSearchResult__panel--sections'>
+          {
+            renderedSections.map((section) => (
+              <MobileSectionPanel
+                key={ section.crn }
+                section={ section }
+                showNotificationSwitches={ userIsWatchingClass }
+              />
+            ))
+          }
+        </div>
+        {!hideShowAll && (
+        <div className='MobileSearchResult__showAll' role='button' tabIndex={ 0 } onClick={ () => setShowAll(!showAll) }>
+          <span>{showAll ? 'Collapse sections' : 'Show all sections'}</span>
+          <IconArrow className={ showAll ? 'MobileSearchResult__showAll--collapse' : '' } />
+        </div>
+        )}
+      </div>
+      )}
+    </div>
+
   )
 }
